Module diem_transaction_builder::stdlib

Conversion library between a structured representation of a Move script call (ScriptCall) and the standard BCS-compatible representation used in Diem transactions (Script).

This code was generated by compiling known Script interfaces (“ABIs”) with the tool transaction-builder-generator.

Enums

ScriptCall
Structured representation of a call into a known Move script.

ScriptFunctionCall
Structured representation of a call into a known Move script function.

Functions

encode_add_currency_to_account_script
Summary

encode_add_currency_to_account_script_function
Summary

encode_add_recovery_rotation_capability_script
Summary

encode_add_recovery_rotation_capability_script_function
Summary

encode_add_validator_and_reconfigure_script
Summary

encode_add_validator_and_reconfigure_script_function
Summary

encode_add_vasp_domain_script_function
Summary

encode_burn_script
Summary

encode_burn_txn_fees_script
Summary

encode_burn_txn_fees_script_function
Summary

encode_burn_with_amount_script_function
Summary

encode_cancel_burn_script
Summary

encode_cancel_burn_with_amount_script_function
Summary

encode_create_child_vasp_account_script
Summary

encode_create_child_vasp_account_script_function
Summary

encode_create_designated_dealer_script
Summary

encode_create_designated_dealer_script_function
Summary

encode_create_parent_vasp_account_script
Summary

encode_create_parent_vasp_account_script_function
Summary

encode_create_recovery_address_script
Summary

encode_create_recovery_address_script_function
Summary

encode_create_validator_account_script
Summary

encode_create_validator_account_script_function
Summary

encode_create_validator_operator_account_script
Summary

encode_create_validator_operator_account_script_function
Summary

encode_create_vasp_domains_script_function
Summary

encode_freeze_account_script
Summary

encode_freeze_account_script_function
Summary

encode_initialize_diem_consensus_config_script_function
Summary

encode_peer_to_peer_by_signers_script_function
Summary

encode_peer_to_peer_with_metadata_script
Summary

encode_peer_to_peer_with_metadata_script_function
Summary

encode_preburn_script
Summary

encode_preburn_script_function
Summary

encode_publish_shared_ed25519_public_key_script
Summary

encode_publish_shared_ed25519_public_key_script_function
Summary

encode_register_validator_config_script
Summary

encode_register_validator_config_script_function
Summary

encode_remove_validator_and_reconfigure_script
Summary

encode_remove_validator_and_reconfigure_script_function
Summary

encode_remove_vasp_domain_script_function
Summary

encode_rotate_authentication_key_script
Summary

encode_rotate_authentication_key_script_function
Summary

encode_rotate_authentication_key_with_nonce_admin_script
Summary

encode_rotate_authentication_key_with_nonce_admin_script_function
Summary

encode_rotate_authentication_key_with_nonce_script
Summary

encode_rotate_authentication_key_with_nonce_script_function
Summary

encode_rotate_authentication_key_with_recovery_address_script
Summary

encode_rotate_authentication_key_with_recovery_address_script_function
Summary

encode_rotate_dual_attestation_info_script
Summary

encode_rotate_dual_attestation_info_script_function
Summary

encode_rotate_shared_ed25519_public_key_script
Summary

encode_rotate_shared_ed25519_public_key_script_function
Summary

encode_set_gas_constants_script_function
Summary

encode_set_validator_config_and_reconfigure_script
Summary

encode_set_validator_config_and_reconfigure_script_function
Summary

encode_set_validator_operator_script
Summary

encode_set_validator_operator_script_function
Summary

encode_set_validator_operator_with_nonce_admin_script
Summary

encode_set_validator_operator_with_nonce_admin_script_function
Summary

encode_tiered_mint_script
Summary

encode_tiered_mint_script_function
Summary

encode_unfreeze_account_script
Summary

encode_unfreeze_account_script_function
Summary

encode_update_diem_consensus_config_script_function
Summary

encode_update_diem_version_script
Summary

encode_update_diem_version_script_function
Summary

encode_update_dual_attestation_limit_script
Summary

encode_update_dual_attestation_limit_script_function
Summary

encode_update_exchange_rate_script
Summary

encode_update_exchange_rate_script_function
Summary

encode_update_minting_ability_script
Summary

encode_update_minting_ability_script_function
Summary