Function diem_transaction_builder::stdlib::encode_remove_vasp_domain_script_function

pub fn encode_remove_vasp_domain_script_function(
    address: AccountAddress, 
    domain: Vec<u8>
) -> TransactionPayload

Remove a VASP domain from parent VASP account. The transaction can only be sent by the Treasury Compliance account.

Technical Description

Removes a VASPDomain::VASPDomain from the domains field of the VASPDomain::VASPDomains resource published under account with address.

Parameters

Name	Type	Description

tc_account	signer	The signer of the sending account of this transaction. Must be the Treasury Compliance account.

address	address	The address of parent VASP account that will update its domains.

domain	vector<u8>	The domain name.

Common Abort Conditions

Error Category	Error Reason	Description

Errors::REQUIRES_ROLE	Roles::ETREASURY_COMPLIANCE	The sending account is not the Treasury Compliance account.

Errors::REQUIRES_ADDRESS	CoreAddresses::ETREASURY_COMPLIANCE	tc_account is not the Treasury Compliance account.

Errors::NOT_PUBLISHED	VASPDomain::EVASP_DOMAIN_MANAGER	The VASPDomain::VASPDomainManager resource is not yet published under the Treasury Compliance account.

Errors::NOT_PUBLISHED	VASPDomain::EVASP_DOMAINS_NOT_PUBLISHED	address does not have a VASPDomain::VASPDomains resource published under it.

Errors::INVALID_ARGUMENT	VASPDomain::EINVALID_VASP_DOMAIN	The domain is greater in length than VASPDomain::DOMAIN_LENGTH.

Errors::INVALID_ARGUMENT	VASPDomain::EVASP_DOMAIN_NOT_FOUND	The domain does not exist in the list of VASPDomain::VASPDomains in the VASPDomain::VASPDomains resource published under address.